Update to version 1.1.1
Added in width option.

Added in new example of width option use along with details about use pertaining to the user experience.

Added in CSS for making the cursor change to a pointer cursor on hover of the reset trigger element.

Tested out tons of hacks for addressing the iOS Safari Input focus bug... none worked. Tested across other desktop and mobile browsers, and found that this is a bug that seems to stem from the iOS 11 Safari client. iOS Chrome and Firefox worked as expected. Will make a statement here on github about the issue just for clarity and transparency for any other devs interensted in using this plugin.